requirements.txt
pandas>=1.5.0
numpy>=1.21.0
geopandas>=0.12.0
osmnx>=1.3.0
networkx>=2.8.0
shapely>=1.8.0
scipy>=1.9.0
requests>=2.28.0
matplotlib>=3.5.0
folium>=0.14.0
r5py>=0.1.0
seaborn>=0.11.0
statsmodels>=0.13.0

# Google Maps API for Distance Matrix
googlemaps>=4.10.0

# Environment variable management
python-dotenv>=0.19.0 